init SidePanel component
