+++ src/app/components/all.scss
@@ -38,7 +38,7 @@
 @import "./modules/SignUp";
 @import "./modules/LoginForm";
 @import "./modules/MiniHeader";
-@import "./modules/SidePanel";
+@import "./modules/SidePanel/styles";
 @import "./modules/Settings";
 @import "./modules/BottomPanel";
 @import "./modules/UserWallet";
