+++ src/app/components/modules/SidePanel/styles.scss
